Refactor: path for constants
