+++ src/pages/Practice/PracticeMain.js
@@ -5,7 +5,7 @@ import styled from "styled-components";
 import PracticeListContent from "../../components/modules/PracticeListContent";
 import Header from "../../components/modules/Header";
 
-import { PRACTICE_LIST } from "../../constants/practice";
+import { PRACTICE_LIST } from "../../common/constants";
 
 function PracticeMain() {
   const navigate = useNavigate();
